Moved validateName from Account/Player to bundle lib
