+++ src/Account.js
@@ -1,7 +1,6 @@
 'use strict';
 const bcrypt = require('bcryptjs');
 const Data   = require('./Data');
-const Config  = require('./Config');
 
 /**
  * Representation of a player's account
@@ -85,29 +84,6 @@ class Account {
   _hashPassword(pass) {
     const salt = bcrypt.genSaltSync(10);
     return bcrypt.hashSync(pass, salt);
-  }
-
-  /**
-   * @param {string} name
-   * @return {boolean}
-   */
-  static validateName(name) {
-    const maxLength = Config.get('maxAccountNameLength');
-    const minLength = Config.get('minAccountNameLength');
-
-    if (!name) {
-      return 'Please enter a name.';
-    }
-    if (name.length > maxLength) {
-      return 'Too long, try a shorter name.';
-    }
-    if (name.length < minLength) {
-      return 'Too short, try a longer name.';
-    }
-    if (!/^[a-z]+$/i.test(name)) {
-      return 'Your name may only contain A-Z without spaces or special characters.';
-    }
-    return false;
   }
 
   /**
